Move clamd client to its own module
